chore(influx_tools): Rename command influx-tools → influx_tools
