+++ cmd/influx_tools/compact/command.go
@@ -13,11 +13,11 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format/binary"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format/line"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/shard"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/storage"
-	"github.com/influxdata/influxdb/cmd/influx-tools/server"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format/binary"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format/line"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/shard"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/storage"
+	"github.com/influxdata/influxdb/cmd/influx_tools/server"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/tsdb"
 	"github.com/influxdata/influxdb/tsdb/engine/tsm1"
