+++ cmd/influx_tools/importer/importer.go
@@ -8,9 +8,9 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/errlist"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/shard"
-	"github.com/influxdata/influxdb/cmd/influx-tools/server"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/errlist"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/shard"
+	"github.com/influxdata/influxdb/cmd/influx_tools/server"
 	"github.com/influxdata/influxdb/services/meta"
 	"github.com/influxdata/influxdb/tsdb"
 	"github.com/influxdata/influxdb/tsdb/engine/tsm1"
