+++ cmd/influx_tools/internal/format/binary/common.go
