+++ cmd/influx_tools/internal/format/binary/messagetype_string.go
