+++ cmd/influx_tools/internal/format/binary/reader.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/tlv"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/tlv"
 	"github.com/influxdata/influxdb/tsdb"
 	"github.com/influxdata/influxdb/tsdb/engine/tsm1"
 )
