+++ cmd/influx_tools/internal/format/binary/reader_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format/binary"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format/binary"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxql"
 )
