+++ cmd/influx_tools/internal/format/binary/writer.go
@@ -6,8 +6,8 @@ import (
 	"io"
 	"time"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/tlv"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/tlv"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/tsdb"
 	"github.com/influxdata/influxql"
