+++ cmd/influx_tools/internal/format/binary/writer_test.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format/binary"
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/tlv"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format/binary"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/tlv"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxql"
 )
