+++ cmd/influx_tools/internal/format/line/writer.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/pkg/escape"
 	"github.com/influxdata/influxdb/tsdb"
