+++ cmd/influx_tools/internal/format/text/writer.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"strconv"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/format"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/format"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/pkg/escape"
 	"github.com/influxdata/influxdb/tsdb"
