+++ cmd/influx_tools/internal/format/writer.go
@@ -3,7 +3,7 @@ package format
 import (
 	"fmt"
 
-	"github.com/influxdata/influxdb/cmd/influx-tools/internal/storage"
+	"github.com/influxdata/influxdb/cmd/influx_tools/internal/storage"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/tsdb"
 	"github.com/influxdata/influxql"
